+++ assets/javascripts/kitten/components/buttons/button-group/button-group.js
@@ -1,0 +1,10 @@
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
+import styled from 'styled-components'
+import { Button } from '../../../components/buttons/button/button'
+
+export class ButtonGroup extends Component {
+  render() {
+    return <Button />
+  }
+}